pyproject.toml
[tool.poetry]
name = "articlesapi"
version = "0.1.0"
description = ""
authors = ["Denis Bobela <youtipie@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.4"
flask = "^3.1.0"
flasgger = "^0.9.7.1"
flask-cors = "^5.0.0"
flask-sqlalchemy = "^3.1.1"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
flask-jwt-extended = "^4.7.1"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
coverage = "^7.6.8"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
